+++ components/browse/Filters.vue
@@ -1,49 +1,50 @@
 <template>
-  <BrowseFilterGroup title="Phân loại" class="pt-4 lg:pt-0">
+  <BrowseFilterGroup :title="$t('common.categories')" class="pt-4 lg:pt-0">
     <ul>
-      <li v-for="(category, key) in categories" :key="key">
-        <NuxtLink :to="{ path: isActive(key) ? '/browse' : `/browse/${key}`, query: getQuery() }"
+      <li v-for="key in sortedCategories" :key="key">
+        <NuxtLinkLocale :to="{ path: isActive(key) ? '/browse' : `/browse/${key}`, query: getQuery() }"
           :class="['flex items-center py-1 space-x-2 transition-colors hover:text-black', isActive(key) ? 'text-black font-semibold' : 'text-gray-500']">
-          <Icon :name="category.icon" size="20" />
-          <span>{{ category.title }}</span>
-        </NuxtLink>
+          <Icon :name="categories[key as CategoryType].icon" size="20" />
+          <span>{{ $t('category.' + key) }}</span>
+        </NuxtLinkLocale>
       </li>
     </ul>
   </BrowseFilterGroup>
-  <BrowseFilterGroup title="Doanh thu">
+  <BrowseFilterGroup :title="$t('common.revenue')">
     <CoreDualSlider :values="selected.revenue.value as number[]" :format='{ prefix: "$", thousand: "," }' :min="0"
       :max="100000" :step="500" @value-commit="handleSlider" />
   </BrowseFilterGroup>
-  <BrowseFilterGroup title="Tình trạng">
+  <BrowseFilterGroup :title="$t('common.status')">
     <ul>
-      <li v-for="(state, key) in status" :key="key">
+      <li v-for="key in status" :key="key">
         <CoreCheckbox :id="prefix(key)" :value="key" name="status" :checked="selected.status.value.includes(key)"
-          @change="handleCheckbox">{{ state }}</CoreCheckbox>
+          @change="handleCheckbox">{{ $t('common.status-types.' + key) }}</CoreCheckbox>
       </li>
     </ul>
   </BrowseFilterGroup>
-  <BrowseFilterGroup title="Mô hình thu nhập">
+  <BrowseFilterGroup :title="$t('common.revenue-models')">
     <ul>
-      <li v-for="(model, key) in revenueModels" :key="key">
+      <li v-for="key in revenueModels" :key="key">
         <CoreCheckbox :id="prefix(key)" :value="key" name="models" :checked="selected.models.value.includes(key)"
-          @change="handleCheckbox">{{ model }}</CoreCheckbox>
+          @change="handleCheckbox">{{ $t('common.revenue-model-types.' + key) }}</CoreCheckbox>
       </li>
     </ul>
   </BrowseFilterGroup>
-  <BrowseFilterGroup title="Được gắn thẻ">
+  <BrowseFilterGroup :title="$t('common.tags')">
     <CoreToggle v-for="tag in tags" :key="tag" :value="tag" class="mr-1 mb-1"
       :pressed="selected.tags.value.includes(tag)" @change="handleToggle">
       {{ tag }}
     </CoreToggle>
   </BrowseFilterGroup>
   <CoreButton v-if="showReset" href="/browse" variant="outline" class="block w-full sticky bottom-4">
-    Reset
+    {{ $t('common.reset') }}
   </CoreButton>
 </template>
 
 <script setup lang="ts">
 type Filter = 'revenue' | 'status' | 'models' | 'tags'
 
+const { t } = useI18n()
 const { isMobile } = defineProps<{ isMobile?: boolean }>()
 const route = useRoute()
 const router = useRouter();
@@ -62,6 +63,7 @@ const selected: Record<Filter, Ref<(string | number)[]>> = {
 }
 const allTags = data.value?.reduce((acc: string[], product) => [...acc, ...product.hashtags], [])
 const tags = Array.from(new Set(allTags))
+const sortedCategories = Object.keys(categories).sort((c1, c2) => t(`category.${c1}`).localeCompare(t(`category.${c2}`)))
 
 const prefix = (id: string) => `${isMobile ? 'm' : 'd'}-${id}`
 
@@ -100,7 +102,7 @@ const handleToggle = (e: { value: string, pressed: boolean }) => {
   pushQuery('tags', result)
 }
 
-const showReset = computed(() => route.path !== '/browse' || Object.keys(route.query).length > 0)
+const showReset = computed(() => !route.path.endsWith('/browse') || Object.keys(route.query).length > 0)
 
 watch(() => route.query, () => {
   Object.keys(selected).forEach((key) => {