+++ components/browse/ProductCard.vue
@@ -1,6 +1,6 @@
 <template>
-  <NuxtLink :href="product._path" :key="product._path" @mouseover="isHovered = true" @mouseleave="isHovered = false"
-    class="flex flex-col">
+  <NuxtLinkLocale :href="product._path" :key="product._path" @mouseover="isHovered = true"
+    @mouseleave="isHovered = false" class="flex flex-col">
     <div class="w-full h-52 overflow-hidden border border-gray-300 relative shrink-0">
       <img v-if="product.image?.src" :src="product.image.src" :alt="product.image.alt"
         class="w-full h-full object-cover" />
@@ -10,15 +10,28 @@
       </div>
     </div>
     <div class="h-full p-3 border border-t-0 border-gray-300">
-      <CoreHeading as="h4">
-        {{ product.title }}
-        <CoreBadge v-if="product.revenue" class="mb-0 text-xs align-top">
-          {{ formatPrice(product.revenue) }}/tháng
-        </CoreBadge>
-      </CoreHeading>
-      <p>{{ product.description }}</p>
+      <CoreHeading as="h4">{{ product.title }}</CoreHeading>
+      <p class="mb-2">{{ product.description }}</p>
+      <div>
+        <CoreTooltip>
+          <template #trigger>
+            <CoreBadge v-if="product.revenue" class="mb-0 align-top">
+              {{ formatPrice(product.revenue, locale) }}/{{ $t('common.month') }}
+            </CoreBadge>
+          </template>
+          {{ $t('common.revenue') }}
+        </CoreTooltip>
+        <CoreTooltip>
+          <template #trigger>
+            <CoreBadge v-if="product.affiliate" class="mb-0 align-top">
+              <Icon name="iconoir:coins" size="20" />
+            </CoreBadge>
+          </template>
+          {{ $t('common.affiliate-program') }}
+        </CoreTooltip>
+      </div>
     </div>
-  </NuxtLink>
+  </NuxtLinkLocale>
 </template>
 
 <script setup lang="ts">
@@ -27,5 +40,6 @@ import type { ParsedContent } from '@nuxt/content/types';
 defineProps<{
   product: ParsedContent | ParsedProduct
 }>()
+const { locale } = useI18n();
 const isHovered = ref(false);
 </script>