+++ components/core/Button.vue
@@ -1,5 +1,5 @@
 <template>
-  <component :is="as || ($attrs.href ? NuxtLink : 'button')" :class="[
+  <component :is="resolvedAs" :href="href" :rel="rel" :class="[
     'rounded text-center transition focus-visible:ring-2 ring-offset-2 ring-black',
     sizes[size],
     styles[$attrs.disabled === '' ? 'disabled' : variant],
@@ -9,19 +9,18 @@
 </template>
 
 <script setup lang="ts">
-const NuxtLink = resolveComponent("NuxtLink");
-
 type Size = 'sm' | "md" | "lg";
 
-type Variant = "outline" | "primary" | "inverted" | "disabled";
+type Variant = "outline" | "primary" | "inverted" | "disabled" | "custom";
 
 interface Props {
-  as?: 'button' | 'a' | 'NuxtLink';
+  as?: 'button' | 'a';
+  href?: string;
   size?: Size;
   variant?: Variant;
 }
 
-withDefaults(defineProps<Props>(), {
+const { as, href } = withDefaults(defineProps<Props>(), {
   size: "lg",
   variant: "primary",
 });
@@ -37,5 +36,9 @@ const styles: Record<Variant, string> = {
   primary: "bg-black text-white hover:bg-gray-800 border-2 border-transparent",
   inverted: "bg-white text-gray-500 border-2 border-transparent hover:text-black",
   disabled: "bg-gray-100 hover:bg-gray-200 border-2 border-transparent",
+  custom: "",
 };
+
+const resolvedAs = as || (href ? resolveLinkComponent(href) : 'button')
+const rel = href && isExternalLink(href) ? 'nofollow noopener noreferrer' : undefined
 </script>