+++ components/landing/Categories.vue
@@ -1,21 +1,21 @@
 <template>
   <h2 class="text-4xl lg:text-5xl text-center font-bold lg:tracking-tight mb-16">
-    Khám phá sản phẩm theo phân loại
+    {{ $t('landing.hot-categories-title') }}
   </h2>
 
   <div class="grid grid-cols-2 lg:grid-cols-4 mb-12 gap-6 lg:gap-12">
-    <NuxtLink v-for="key in categoryKeys" :href="'/browse/' + key"
+    <NuxtLinkLocale v-for="key in categoryKeys" :href="'/browse/' + key"
       class="text-center text-gray-500 border rounded-sm px-4 py-8 transition-colors hover:border-gray-400 hover:text-black">
       <div class="inline-block p-2 bg-gray-200 rounded-full">
         <Icon :name="categories[key].icon" size="24" />
       </div>
-      <h3 class="font-semibold lg:text-lg">{{ categories[key].title }}</h3>
-    </NuxtLink>
+      <h3 class="font-semibold lg:text-lg">{{ $t('category.' + key) }}</h3>
+    </NuxtLinkLocale>
   </div>
 
   <div class="text-center mb-40">
     <CoreButton href="/browse" class="inline-flex items-center">
-      <span>Xem tất cả</span>
+      <span>{{ $t('landing.view-all') }}</span>
       <Icon name="uil:arrow-right" class="ml-2" size="24" />
     </CoreButton>
   </div>