+++ components/landing/NewProducts.vue
@@ -1,20 +1,28 @@
 <template>
   <h2 class="text-4xl lg:text-5xl text-center font-bold lg:tracking-tight mb-16">
-    Sản phẩm được thêm gần đây
+    {{ $t('landing.new-products-title') }}
   </h2>
 
-  <div class="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-4 gap-6 mb-12">
-    <BrowseProductCard v-for="product in data" :key="product._path" :product="product" />
-  </div>
+  <LazyContentList path="/product" :query="query" v-slot="{ list }">
+    <div class="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-4 gap-6 mb-12">
+      <BrowseProductCard v-for="product in list" :key="product._path" :product="product" />
+    </div>
+  </LazyContentList>
 
   <div class="text-center mb-40">
     <CoreButton href="/browse" class="inline-flex items-center">
-      <span>Xem tất cả</span>
+      <span>{{ $t('landing.view-all') }}</span>
       <Icon name="uil:arrow-right" class="ml-2" size="24" />
     </CoreButton>
   </div>
 </template>
 
 <script setup lang="ts">
-const { data } = await useLatestProductsQuery()
+const { locale } = useI18n()
+const query = computed(() => ({
+  where: [{ _locale: locale.value }],
+  without: ['body', 'hashtags'],
+  limit: 4,
+  sort: [{ publishedAt: -1 }]
+}))
 </script>