+++ components/layout/Navbar.vue
@@ -2,12 +2,12 @@
   <header class="w-full">
     <div class="max-w-screen-xl flex flex-col lg:flex-row justify-between items-center p-5 mx-auto">
       <div class="flex w-full lg:w-auto items-center justify-between">
-        <NuxtLink href="/" class="flex items-center text-lg" @mouseover="titleMouseOver = true"
+        <NuxtLinkLocale href="/" class="flex items-center text-lg" @mouseover="titleMouseOver = true"
           @mouseout="titleMouseOver = false">
           <Icon :name="'healthicons:money-bag' + (titleMouseOver ? '' : '-outline')" size="32" class="mr-1" />
           <span class="font-bold text-slate-800">Thu Nhập</span>
           <span class="text-slate-500">.online</span>
-        </NuxtLink>
+        </NuxtLinkLocale>
         <button @click="openMenu = !openMenu"
           class="block p-1 lg:hidden text-gray-500 rounded-full transition-colors hover:text-black hover:bg-gray-100">
           <svg fill="currentColor" class="w-6 h-6" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
@@ -23,32 +23,43 @@
       </div>
       <nav class="w-full items-center lg:w-auto mt-2 lg:flex lg:mt-0" :class="{ block: openMenu, hidden: !openMenu }">
         <LayoutSearch />
-        <ul class="flex flex-col lg:flex-row lg:gap-3 mb-2 lg:mb-0 lg:mr-3">
-          <li v-for="item of menuitems ">
-            <NuxtLink :href="item.path"
-              :class='[($route.path.includes(item.path) ? "text-black bg-gray-100" : "text-gray-500"), "rounded-md flex p-2 transition-colors hover:text-black lg:px-3"]'>
-              {{ item.title }}
-            </NuxtLink>
+        <ul class="flex flex-col lg:flex-row lg:gap-3 lg:mr-3 mb-2 lg:mb-0">
+          <li v-for="item of menuitems">
+            <NuxtLinkLocale :href="'/' + item"
+              :class='[($route.path.includes(item) ? "text-black bg-gray-100" : "text-gray-500"), "rounded-md flex p-2 transition-colors hover:text-black lg:px-3"]'>
+              {{ $t('common.' + item) }}
+            </NuxtLinkLocale>
           </li>
         </ul>
+        <div class="flex items-center">
+          <CoreToggleGroup v-model="currentLocale"
+            :items="locales.map((({ code }) => ({ label: code.toUpperCase(), value: code })))" />
+          <NuxtLink :href="contactLinks.twitter.url" target="_blank"
+            class="ml-3 p-2 transition-colors text-gray-500 hover:text-black">
+            <Icon name="uil:twitter" size="20" />
+          </NuxtLink>
+          <NuxtLink :href="contactLinks.email.url" class="p-2 transition-colors text-gray-500 hover:text-black">
+            <Icon name="uil:envelope-alt" size="20" />
+          </NuxtLink>
+        </div>
         <CoreButton href="/submit" variant="primary" size="md" class="inline-block">Đăng sản phẩm</CoreButton>
+        <!-- <div class="lg:hidden flex items-center mt-3 gap-4">
+          <LandingLink href="#" styleName="muted" block size="md"
+            >Log in</LandingLink
+          >
+          <LandingLink href="#" size="md" block>Sign up</LandingLink>
+        </div> -->
       </nav>
     </div>
   </header>
 </template>
 
 <script setup lang="ts">
-const menuitems = [
-  {
-    title: "Khám phá",
-    path: "/browse",
-  },
-  {
-    title: "Giới thiệu",
-    path: "/about",
-  },
-];
-
+const { locale, locales, setLocale } = useI18n()
 const openMenu = ref(false);
 const titleMouseOver = ref(false);
+const currentLocale = ref(locale.value)
+const menuitems = ['browse', 'about']
+
+watch(currentLocale, setLocale)
 </script>