+++ nuxt.config.ts
@@ -1,3 +1,10 @@
+const baseUrl = "https://thunhap.online";
+const defaultLocale = "en";
+const locales = [
+  { code: "vi", iso: "vi-VN", file: "vi.json" },
+  { code: "en", iso: "en-US", file: "en.json" },
+];
+
 // https://nuxt.com/docs/api/configuration/nuxt-config
 export default defineNuxtConfig({
   devtools: { enabled: true },
@@ -9,6 +16,8 @@ export default defineNuxtConfig({
     },
   },
   content: {
+    locales: locales.map((locale) => locale.code),
+    defaultLocale,
     experimental: {
       search: {},
     },
@@ -16,12 +25,14 @@ export default defineNuxtConfig({
   modules: [
     "@nuxt/content",
     "@nuxtjs/seo",
+    "@nuxtjs/i18n",
+    "@nuxthq/studio",
     "@vueuse/nuxt",
     "nuxt-icon",
     "nuxt-gtag",
   ],
   site: {
-    url: "https://thunhap.online",
+    url: baseUrl,
     name: "Thu Nhập Online",
     description: "Khám phá các sản phẩm online thành công, mang lại thu nhập",
   },
@@ -31,5 +42,11 @@ export default defineNuxtConfig({
   },
   ogImage: {
     enabled: false,
+  },
+  i18n: {
+    baseUrl,
+    langDir: "locales",
+    locales,
+    defaultLocale,
   },
 });