+++ server/api/__sitemap__/urls.ts
@@ -8,11 +8,12 @@ export default defineSitemapEventHandler(async (e) => {
     asSitemapUrl({
       loc: c._path,
       images: c.image ? [{ loc: c.image.src, title: c.image.alt }] : undefined,
+      _i18nTransform: true,
     })
   );
 
   const categoryUrls = Object.keys(categories).map((category) =>
-    asSitemapUrl({ loc: `/browse/${category}` })
+    asSitemapUrl({ loc: `/browse/${category}`, _i18nTransform: true })
   );
 
   return [...categoryUrls, ...contentUrls];