+++ utils/filters.ts
@@ -1,73 +1,61 @@
-export const categories = {
+import type { CategoryType, RevenueModelType, StatusType } from "./types";
+
+export const categories: Record<CategoryType, any> = {
   ai: {
-    title: "AI",
     icon: "uil:robot",
   },
   automation: {
-    title: "Tự động hóa",
     icon: "uil:cog",
   },
   marketing: {
-    title: "Marketing",
     icon: "uil:megaphone",
   },
   design: {
-    title: "Thiết kế",
     icon: "uil:palette",
   },
   website: {
-    title: "Công cụ làm web",
     icon: "uil:web-grid",
   },
   form: {
-    title: "Công cụ tạo form",
     icon: "uil:file-plus-alt",
   },
   mobile: {
-    title: "Di động",
     icon: "uil:mobile-android",
   },
   analytics: {
-    title: "Phân tích dữ liệu",
     icon: "uil:chart-pie",
   },
   editor: {
-    title: "Chỉnh sửa ảnh/video",
     icon: "uil:image-edit",
   },
   template: {
-    title: "Template",
     icon: "uil:box",
   },
   utility: {
-    title: "Tiện ích",
     icon: "uil:wrench",
   },
   service: {
-    title: "Dịch vụ",
     icon: "uil:calling",
   },
   entertainment: {
-    title: "Giải trí",
     icon: "uil:play-circle",
   },
   jobs: {
-    title: "Trang tìm việc",
     icon: "uil:briefcase-alt",
   },
 };
 
-export const revenueModels = {
-  subscription: "Thu phí định kỳ",
-  advertising: "Tiền quảng cáo",
-  transaction: "Giao dịch",
-  sponsorship: "Tài trợ",
-  affiliate: "Tiếp thị liên kết",
-};
+export const revenueModels: RevenueModelType[] = [
+  "subscription",
+  "advertising",
+  "transaction",
+  "sponsorship",
+  "affiliate",
+];
 
-export const status = {
-  active: "Hoạt động",
-  inactive: "Đã ngừng",
-  acquired: "Đã bán",
-  unknown: "Không rõ",
-};
+export const status: StatusType[] = [
+  "active",
+  "inactive",
+  "acquired",
+  "unknown",
+];