+++ utils/links.ts
@@ -1,0 +1,16 @@
+export const contactLinks = {
+  email: {
+    text: "hello@thunhap.online",
+    url: "mailto:hello@thunhap.online",
+  },
+  twitter: {
+    text: "thunhap_online",
+    url: "https://twitter.com/thunhap_online",
+  },
+};
+
+export const affiliateLinks = {
+  "compressx.app": "https://compressx.app/?aff=Gmbpn",
+  "typingmind.com": "",
+  "carrd.co": "https://try.carrd.co/rhb9ffjr",
+};