+++ utils/misc.ts
@@ -1,3 +1,5 @@
+import { NuxtLink, NuxtLinkLocale } from "#components";
+
 export const countElements = <T>(array: T[]): { [key: string]: number } => {
   let counts: { [key: string]: number } = {};
   array.forEach((element) => {
@@ -6,3 +8,8 @@ export const countElements = <T>(array: T[]): { [key: string]: number } => {
   });
   return counts;
 };
+
+export const isExternalLink = (href: string) => /^(https?|mailto)/.test(href);
+
+export const resolveLinkComponent = (href: string) =>
+  isExternalLink(href) ? NuxtLink : NuxtLinkLocale;