+++ content/vi/changelog.md
@@ -2,6 +2,24 @@
 
 Các cập nhật và sửa đổi đã thực hiện trên hệ thống trang web
 
+## 12-04-2024
+
+- Tinh chỉnh bố cục phần Tổng quan sản phẩm
+- Thêm thông tin "Chương trình liên kết" cho sản phẩm
+
+## 07-04-2024
+
+- Thêm "thông tin trang web" vào phần Tổng quan sản phẩm
+
+## 06-04-2024
+
+- Hiển thị đơn vị tiền ứng với ngôn ngữ
+- Thêm Twitter vào footer
+
+## 03-04-2024
+
+- Thêm tiếng Anh cho trang web
+
 ## 31-03-2024
 
 - Thêm chức năng phân trang cho danh sách sản phẩm